pyproject.toml
[build-system]
requires = ["setuptools>=67"]
build-backend = "setuptools.build_meta"

[project]
name = "green-agent-benchmark"
version = "0.1.0"
description = "Reference implementation of the Green Agent Benchmark for Texas Hold'em evaluation."
requires-python = ">=3.10"
dependencies = [
  "pyyaml>=6.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4"
]

[tool.setuptools.packages.find]
where = ["."]
